Rename Truncate(scope) to Clear(scope)
There was no reason to introduce another wording for a command that
already exists. So, truncating a scope is basically clearing a scope. As
there is already an implementation to clear an entry clear(scope, key)),
it is clever to utilize c++ overloading feature to just create the same
command but only passing a scooe.