Refactor Keeper to be defined with KeepCommand
By moving KeepCommand and KeeperEntry to the same fine, which mean that
the command knows more closely about the implementation details of the
entry it interacts with, there is less need to have interfaces defined.
For example, there is no need now to have a KeeperEntry interface which
then reduces complexity and better respects Open and Close Principle.